week7/gradio-oxford-pet-app/requirements.txt
transformers
torch
openai
python-dotenv